IE10 submit fix
In IE10, the click event fired by the script once validation was passed
would not trigger the submit event for the form. It now uses a
duplicated button that overlays the registration button. So the user
clicks the registration button and then the script triggers the click on
the register button (which now works because no click event handlers are
bound to it).